Refactor ExtendedEda with helpers
